+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.ATSOPTANDROID"
         tools:targetApi="31">
         <activity
-            android:name=".splash.SplashActivity"
+            android:name=".ui.splash.SplashActivity"
             android:exported="true"
             android:theme="@style/Theme.ATSOPTANDROID" >
             <intent-filter>
@@ -23,15 +23,15 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".my.MyActivity"
+            android:name=".ui.my.MyActivity"
             android:exported="false"
             android:theme="@style/Theme.ATSOPTANDROID" />
         <activity
-            android:name=".login.LoginActivity"
+            android:name=".ui.login.LoginActivity"
             android:exported="false"
             android:theme="@style/Theme.ATSOPTANDROID" />
         <activity
-            android:name=".signup.SignUpActivity"
+            android:name=".ui.signup.SignUpActivity"
             android:exported="false"
             android:theme="@style/Theme.ATSOPTANDROID" />
     </application>
