+++ app/src/main/java/org/sopt/at/ui/my/MyActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.my
+package org.sopt.at.ui.my
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,17 +6,11 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -24,9 +18,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import org.sopt.at.R
-import org.sopt.at.splash.SplashActivity
-import org.sopt.at.ui.HeaderLayout
+import org.sopt.at.ui.my.component.LogoutButton
+import org.sopt.at.ui.splash.SplashActivity
+import org.sopt.at.custom.HeaderLayout
 import org.sopt.at.util.MyApplication.Companion.PREFS_ID_KEY
 import org.sopt.at.util.MyApplication.Companion.prefs
 
@@ -60,26 +54,7 @@ class MyActivity : ComponentActivity() {
                 color = Color.White
             )
             Spacer(modifier = Modifier.weight(1f))
-            Button(
-                onClick = {
-                    logout()
-                },
-                shape = RoundedCornerShape(2.dp),
-                colors = ButtonDefaults.textButtonColors(
-                    containerColor = Color.Black,
-                    disabledContainerColor = Color.Gray,
-
-                    contentColor = Color.White,
-                    disabledContentColor = Color.LightGray
-                ),
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(48.dp)
-                    .border(1.dp, Color.Gray),
-                content = {
-                    Text(text = resources.getString(R.string.my_logout_btn))
-                }
-            )
+            LogoutButton(logoutOnClick = { logout() })
         }
     }
