+++ app/src/main/java/org/sopt/at/ui/signup/navigation/SignUpScreenRoute.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.signup.navigation
+package org.sopt.at.ui.signup.navigation
 
 sealed class SignUpScreenRoute(val route: String) {
     data object SignUpIdScreen: SignUpScreenRoute("SignupIdScreen")
