+++ app/src/main/java/org/sopt/at/ui/signup/screen/SignUpPwScreen.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.signup.screen
+package org.sopt.at.ui.signup.screen
 
 import android.widget.Toast
 import androidx.compose.foundation.background
@@ -10,10 +10,13 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TextFieldDefaults
@@ -26,25 +29,25 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import org.sopt.at.R
-import org.sopt.at.signup.navigation.SignUpScreenRoute
+
 
 @Composable
-fun SignUpIdScreen(
-    navController: NavHostController,
-    idCallback: (String) -> Unit
+fun SignUpPwScreen(
+    signUpAccountState: SignUpAccountState,
+    signUpEndCallback: () -> Unit
 ) {
     val context = LocalContext.current
-    var idTextValue by remember { mutableStateOf("") }
-    var isNextBtnEnable by remember { mutableStateOf(false) }
-
-    isNextBtnEnable = idTextValue.isNotEmpty()
+    var shouldShowPassword by remember { mutableStateOf(false) }
 
     Column (
         modifier = Modifier
@@ -56,7 +59,7 @@ fun SignUpIdScreen(
 
         // Title
         Text(
-            text = context.resources.getString(R.string.signup_id_title),
+            text = context.resources.getString(R.string.signup_pw_title),
             fontSize = 24.sp,
             color = Color.White,
             modifier = Modifier.align(Alignment.CenterHorizontally)
@@ -64,11 +67,11 @@ fun SignUpIdScreen(
 
         Spacer(modifier = Modifier.height(24.dp))
 
-        // ID TextField
+        // Pw TextField
         TextField(
-            value = idTextValue,
-            onValueChange = { idTextValue = it },
-            placeholder = { Text(text = context.resources.getString(R.string.signup_id_placeholder)) },
+            value = signUpAccountState.pw,
+            onValueChange = { signUpAccountState.pw = it },
+            placeholder = { Text(text = context.resources.getString(R.string.signup_pw_placeholder)) },
             singleLine = true,
             shape = RoundedCornerShape(2.dp),
             textStyle = TextStyle(
@@ -91,6 +94,23 @@ fun SignUpIdScreen(
                 focusedPlaceholderColor = Color.Gray,
                 unfocusedPlaceholderColor = Color.Gray
             ),
+            trailingIcon = {
+                IconButton(
+                    onClick = {
+                        shouldShowPassword = !shouldShowPassword
+                    },
+                    modifier = Modifier.size(24.dp)
+                ) {
+                    Icon(
+                        painter = painterResource(id = if(shouldShowPassword) R.drawable.icon_eye_open else R.drawable.icon_eye_close),
+                        tint = Color.White,
+                        contentDescription = "Password Toggle"
+                    )
+                }
+            },
+            visualTransformation =
+                if(shouldShowPassword) VisualTransformation.None
+                else PasswordVisualTransformation(),
             keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Done),
             modifier = Modifier
                 .fillMaxWidth()
@@ -100,7 +120,7 @@ fun SignUpIdScreen(
         Spacer(modifier = Modifier.height(16.dp))
 
         Text(
-            text = context.resources.getString(R.string.signup_id_condition),
+            text = context.resources.getString(R.string.signup_pw_condition),
             fontSize = 12.sp,
             color = Color.Gray,
         )
@@ -110,13 +130,12 @@ fun SignUpIdScreen(
         // Next Btn
         Button(
             onClick = {
-                if(isIdValid(idTextValue)) {
-                    navController.navigate(SignUpScreenRoute.SignUpPwScreen.route)
-                    idCallback.invoke(idTextValue)
+                if(isPwValid(signUpAccountState.pw)) {
+                    signUpEndCallback()
                 }
-                else Toast.makeText(context, context.resources.getString(R.string.toast_signup_id_fail), Toast.LENGTH_SHORT).show()
+                else Toast.makeText(context, context.resources.getString(R.string.toast_signup_pw_fail), Toast.LENGTH_SHORT).show()
             },
-            enabled = isNextBtnEnable,
+            enabled = signUpAccountState.isNextBtnEnableForPw,
             shape = RoundedCornerShape(2.dp),
             colors = ButtonDefaults.textButtonColors(
                 containerColor = Color.Black,
@@ -131,7 +150,7 @@ fun SignUpIdScreen(
                 .border(1.dp, Color.Gray),
             content = {
                 Text(
-                    text = context.resources.getString(R.string.signup_id_next_btn),
+                    text = context.resources.getString(R.string.signup_pw_next_btn),
                     fontSize = 16.sp,
                     fontWeight = FontWeight.Bold,
                 )
@@ -142,7 +161,7 @@ fun SignUpIdScreen(
     }
 }
 
-fun isIdValid(input: String): Boolean {
-    val regex = "^[a-zA-Z0-9]{6,12}$".toRegex()
+fun isPwValid(input: String): Boolean {
+    val regex = "^[a-zA-Z0-9~!@#$%^&*]{8,15}$".toRegex()
     return regex.matches(input)
 }
