+++ app/src/main/java/org/sopt/at/ui/splash/SplashActivity.kt
@@ -1,10 +1,8 @@
-package org.sopt.at.splash
+package org.sopt.at.ui.splash
 
 import android.annotation.SuppressLint
 import android.content.Intent
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
@@ -15,44 +13,45 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
+import kotlinx.coroutines.delay
 import org.sopt.at.R
-import org.sopt.at.login.LoginActivity
-import org.sopt.at.my.MyActivity
+import org.sopt.at.ui.login.LoginActivity
+import org.sopt.at.ui.my.MyActivity
 import org.sopt.at.util.MyApplication.Companion.PREFS_ID_KEY
 import org.sopt.at.util.MyApplication.Companion.prefs
 
 @SuppressLint("CustomSplashScreen")
 class SplashActivity : ComponentActivity() {
 
-    private var splashHandler = Handler(Looper.getMainLooper())
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
+            LaunchedEffect(Unit) {
+                delay(2000)
+                val insertedId = prefs.getData(PREFS_ID_KEY)
+
+                val intent = if(insertedId.isNullOrBlank()) {
+                    Intent(applicationContext, LoginActivity::class.java)
+                } else {
+                    Intent(applicationContext, MyActivity::class.java)
+                }
+                startActivity(intent)
+                finish()
+            }
+
             Scaffold(modifier = Modifier
                 .fillMaxSize()
                 .background(Color.Black)
             ) { innerPadding ->
                 Box(modifier = Modifier.padding(innerPadding)) {
                     ContentLayout()
-
-                    splashHandler.postDelayed({
-                        val insertedId = prefs.getData(PREFS_ID_KEY)
-
-                        val intent = if(insertedId.isNullOrBlank()) {
-                            Intent(applicationContext, LoginActivity::class.java)
-                        } else {
-                            Intent(applicationContext, MyActivity::class.java)
-                        }
-                        startActivity(intent)
-                        finish()
-                    }, 2000)
                 }
             }
         }
